Update ReleaseNotes2.3.0OpenSource.md
Added updated link for Slack channels, Which will redirect the user to the Slack Application instead of opening the Slack channel in the browser. If the user is not having a Slack Application, then Slack channels will open in the browser. The purpose is to gain more traffic on the Slack Channel.